Separate services in modules, add cron to services.
