+++ currency_rate_update/__openerp__.py
@@ -87,7 +87,7 @@ Thanks to main contributors: Grzegorz Grzelak, Alexis de Lattre
         "account",  # Added to ensure account security groups are present
     ],
     "data": [
-        #"company_data.xml",
+        "service_cron_data.xml",
         "currency_rate_update.xml",
         "company_view.xml",
         "security/security.xml",
